Fix broken concurrent tests
This commit fixes a bug where tests would hang if not run in serial mode.
(ava --serial)

AVA test files are run in parallel, and individual tests are concurrent.
Many tests defined in one file will run concurrently, but one test in many
files will run in parallel. To know the difference, check reference 1.

Going forward, all tests that deal with different fixtures should be placed
in different files, and any time you would normally use `describe` in other
test suites, create a new AVA test file.

References:
1: http://stackoverflow.com/questions/1050222/concurrency-vs-parallelism-what-is-the-difference
